Rename folder and tenant
