+++ front/app/containers/SignUpInContainer/utils/openSignUpInModalIfNecessary.ts
